Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 int main(){
-    int a;
+    float a;
     float pi=3.14,area;
-    scanf("%d",&a);
+    scanf("%f",&a);
     area=pi*a;
     printf("Area: %.2f",area);
     return 0;
